--- package.json
@@ -1,6 +1,0 @@
-{
-  "dependencies": {
-    "bootstrap": "^5.2.1",
-    "react-bootstrap": "^2.5.0"
-  }
-}
