--- raaghu-elements/src/components/index.ts
@@ -1,6 +1,0 @@
-export {default as RdsButton} from './rds-button';
-export {default as RdsButtonGroup} from './rds-button-group';
-export {default as RdsInput} from './rds-input';
-export {default as RdsCheckbox} from './rds-checkbox';
-export {default as RdsCheckboxParentChild} from './rds-checkbox-parent-child';
-export {default as RdsAvatar} from './rds-avatar';
